+++ src/dawn_wrapper.h
@@ -28,8 +28,8 @@ struct buffer_wrapper {
     buffer_wrapper& write(const std::vector<uint8_t>& colors);
     buffer_wrapper& write(const void*);
     bool done();
-    buffer_wrapper& get_output(std::function<void(unsigned, const void*)>);
-    unsigned long get_size();
+    buffer_wrapper& get_output(std::function<void(size_t, const void*)>);
+    size_t get_size();
 
     operator bool() const;
     WRAPPER_PIMPL_DEC(buffer_wrapper);
@@ -122,12 +122,12 @@ struct dawn_plugin {
     ~dawn_plugin();
     render_wrapper make_render();
     compute_wrapper make_compute();
-    buffer_wrapper make_src_buffer(unsigned size, buffer_type type);
-    buffer_wrapper make_dst_buffer(unsigned size, buffer_type type);
-    texture_wrapper make_texture(unsigned);
-    texture_wrapper make_texture(unsigned, unsigned);
+    buffer_wrapper make_src_buffer(size_t size, buffer_type type);
+    buffer_wrapper make_dst_buffer(size_t size, buffer_type type);
+    texture_wrapper make_texture(size_t);
+    texture_wrapper make_texture(size_t, size_t);
     texture_wrapper make_texture(std::vector<uint8_t> data);
-    texture_output_wrapper make_texture_output(unsigned, unsigned);
+    texture_output_wrapper make_texture_output(size_t, size_t);
     encoder_wrapper make_encoder();
     bool run();
     operator bool() const;
